update(pkg): refactor pkg.sh and add fd.env

+++ pkg/pkg.sh
@@ -5,32 +5,33 @@ APP=$1
 PREFIX=.
 
 # # Environment variables template file `${APP}.env`
-# APP_BIN=
-# APP_VERSION=
-# APP_HASH=
-# APP_HASH_URL=
-# APP_PKG=
-# APP_PKG_SRC=
-# APP_PKG_SRC_BIN=
-# APP_PKG_SRC_COMP=
-# APP_PKG_SRC_MAN=
-# APP_PKG_TYPE=[APT | BIN | DEB | TARGZ | TARXZ ]
-# APP_PKG_URL=
+# export APP_BIN=
+# export APP_VERSION=
+# export APP_HASH=
+# export APP_HASH_URL=
+# export APP_PKG=
+# export APP_PKG_SRC=
+# export APP_PKG_SRC_BIN=
+# export APP_PKG_SRC_COMP=
+# export APP_PKG_SRC_MAN=
+# export APP_PKG_TYPE=[APT | BIN | DEB | TARGZ | TARXZ ]
+# export APP_PKG_URL=
 
-if [ -f "$(dirname $0)/${APP}.env" ]; then
-  source "$(dirname $0)/${APP}.env"
+if [ -f "$(dirname "$0")/${APP}.env" ]; then
+  # shellcheck source=./example.env
+  source "$(dirname "$0")/${APP}.env"
 else
-  echo "Error: $(dirname $0)/${APP}.env not found"
+  echo "Error: $(dirname "$0")/${APP}.env not found"
   exit 1
 fi
 
-APP_INSTALLED=$(command -v ${APP_BIN} 2>&1 >/dev/null && echo "1" || echo "0")
+APP_INSTALLED=$(command -v "${APP_BIN}" 2>&1 >/dev/null && echo "1" || echo "0")
 
 # Functions
 
 download() {
   if [ "${PREFIX}/${APP_PKG}" != "${PREFIX}/" ]; then
-    wget -qO ${PREFIX}/${APP_PKG} ${APP_PKG_URL}
+    wget -qO "${PREFIX}/${APP_PKG}" "${APP_PKG_URL}"
   fi
 }
 
@@ -41,9 +42,9 @@ extract() {
     ;;
   TARGZ | TARXZ)
     if [ -z "${APP_PKG_SRC}" ]; then
-      tar -C ${PREFIX} -xf ${PREFIX}/${APP_PKG} ${APP_BIN}
+      tar -C ${PREFIX} -xf "${PREFIX}/${APP_PKG}" "${APP_BIN}"
     else
-      tar -C ${PREFIX} -xf ${PREFIX}/${APP_PKG}
+      tar -C ${PREFIX} -xf "${PREFIX}/${APP_PKG}"
     fi
     ;;
   *)
@@ -56,35 +57,35 @@ checksum() {
   if [ "${PREFIX}/${APP_HASH}" != "${PREFIX}/" ] && [ -f "${PREFIX}/${APP_HASH}" ]; then
     echo "Skipping download. ${APP_HASH} already exists."
   elif [ "${PREFIX}/${APP_HASH}" != "${PREFIX}/" ]; then
-    wget -qO ${PREFIX}/${APP_HASH} ${APP_HASH_URL}
+    wget -qO "${PREFIX}/${APP_HASH}" "${APP_HASH_URL}"
   fi
-  sha256sum -c ${APP_HASH}
+  sha256sum -c "${APP_HASH}"
 }
 
-install() {
+install_pkg() {
   case "${APP_PKG_TYPE}" in
   APT)
     sudo apt update
-    sudo apt install -y ${APP_BIN}
+    sudo apt install -y "${APP_BIN}"
     ;;
   DEB)
-    sudo dpkg -i ${PREFIX}/${APP_PKG}
+    sudo dpkg -i "${PREFIX}/${APP_PKG}"
     ;;
   BIN | TARGZ | TARXZ)
     if [ "${PREFIX}/${APP_BIN}" != "${PREFIX}/" ] && [ -f "${PREFIX}/${APP_BIN}" ]; then
-      sudo install -D -m 755 ${PREFIX}/${APP_BIN} /usr/local/bin/${APP_BIN}
+      sudo install -D -m 755 "${PREFIX}/${APP_BIN}" "/usr/local/bin/${APP_BIN}"
     elif [ "${PREFIX}/${APP_PKG_SRC_BIN}" != "${PREFIX}/" ] && [ -d "${PREFIX}/${APP_PKG_SRC_BIN}" ]; then
-      sudo install -D -m 755 ${PREFIX}/${APP_PKG_SRC_BIN}/${APP_BIN} /usr/local/bin/${APP_BIN}
+      sudo install -D -m 755 "${PREFIX}/${APP_PKG_SRC_BIN}/${APP_BIN}" "/usr/local/bin/${APP_BIN}"
     else
       echo "Installation failed. ${APP_BIN} not found."
       exit 1
     fi
     if [ "${PREFIX}/${APP_PKG_SRC_COMP}" != "${PREFIX}/" ] && [ -f "${PREFIX}/${APP_PKG_SRC_COMP}" ]; then
-      sudo cp ${PREFIX}/${APP_PKG_SRC_COMP} /usr/local/share/zsh/site-functions
+      sudo cp "${PREFIX}/${APP_PKG_SRC_COMP}" /usr/local/share/zsh/site-functions
     fi
     if [ "${PREFIX}/${APP_PKG_SRC_MAN}" != "${PREFIX}/" ] && [ -d "${PREFIX}/${APP_PKG_SRC_MAN}" ]; then
       sudo mkdir -p /usr/local/share/man/man1
-      sudo rsync -av ${PREFIX}/${APP_PKG_SRC_MAN}/ /usr/local/share/man/man1
+      sudo rsync -av "${PREFIX}/${APP_PKG_SRC_MAN}/" /usr/local/share/man/man1
     fi
     ;;
   *)
@@ -96,28 +97,28 @@ install() {
 cleanup() {
   if [ "${PREFIX}/${APP_BIN}" != "${PREFIX}/" ] && [ -f "${PREFIX}/${APP_BIN}" ]; then
     echo "Cleaning up ${PREFIX}/${APP_BIN}..."
-    rm ${PREFIX}/${APP_BIN}
+    rm "${PREFIX}/${APP_BIN}"
   fi
 
   if [ "${PREFIX}/${APP_HASH}" != "${PREFIX}/" ] && [ -f "${PREFIX}/${APP_HASH}" ]; then
     echo "Cleaning up ${PREFIX}/${APP_HASH}..."
-    rm ${PREFIX}/${APP_HASH}
+    rm "${PREFIX}/${APP_HASH}"
   fi
 
   if [ "${PREFIX}/${APP_PKG}" != "${PREFIX}/" ] && [ -f "${PREFIX}/${APP_PKG}" ]; then
     echo "Cleaning up ${PREFIX}/${APP_PKG}..."
-    rm ${PREFIX}/${APP_PKG}
+    rm "${PREFIX}/${APP_PKG}"
   fi
 
   if [ "${PREFIX}/${APP_PKG_SRC}" != "${PREFIX}/" ] && [ -d "${PREFIX}/${APP_PKG_SRC}" ]; then
     echo "Cleaning up ${PREFIX}/${APP_PKG_SRC}..."
-    rm -fr ${PREFIX}/${APP_PKG_SRC}
+    rm -fr "${PREFIX:?}/${APP_PKG_SRC}"
   fi
 }
 
 # Main program
 
-if [ ${APP_INSTALLED} -gt 0 ]; then
+if [ "${APP_INSTALLED}" -gt 0 ]; then
   echo "${APP_BIN} is already installed"
 else
   if [ -z "${APP_PKG}" ] || [ -z "${APP_PKG_URL}" ]; then
@@ -148,7 +149,7 @@ else
   fi
 
   echo "Installing ${APP_BIN}..."
-  install
+  install_pkg
 
   # Remove temporary files
   cleanup
